pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-orchestration"
version = "1.0.0"
description = "Orchestration system for AI coding agents across computational experiments"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pandas>=1.3.0",
    "requests>=2.25.0",
    "pyyaml>=5.4.0",
    "python-slugify>=5.0.0",
    "jinja2>=3.0.0",
    "python-dotenv>=0.19.0",
    "openpyxl>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/nickharris808/zero-shot-ai-agents"
Documentation = "https://github.com/nickharris808/zero-shot-ai-agents/tree/main/docs"
Repository = "https://github.com/nickharris808/zero-shot-ai-agents"
Issues = "https://github.com/nickharris808/zero-shot-ai-agents/issues"

[project.scripts]
jules-orchestrator = "providers.jules.orchestrator:main"
openhands-orchestrator = "providers.openhands.orchestrator:main"
augment-orchestrator = "providers.augment.orchestrator:main"
cosine-orchestrator = "providers.cosine.orchestrator:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["providers*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
